+++ Code/Input.h
@@ -18,10 +18,10 @@ struct ControllerState
 {
 	ControllerState();
 	void Set( const ControllerState &state );
-	double leftStickMagnitude; // 0 - 1.0
-	double leftStickRadians;
-	double rightStickMagnitude; // 0 - 1.0
-	double rightStickRadians;
+	float leftStickMagnitude; // 0 - 1.f
+	float leftStickRadians;
+	float rightStickMagnitude; // 0 - 1.f
+	float rightStickRadians;
 	BYTE leftTrigger;
 	BYTE rightTrigger;
 	bool start;
